+++ CHANGELOG.md
@@ -5,6 +5,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Changed
+
+- `winapi` is no longer required on unix; `nix` is no longer required on windows.
+
 ## [0.9.0] - 2020-07-01
 
 ### Added